fixes #199

+++ example/test/descriptor_test_gtest.cpp
@@ -79,14 +79,14 @@ TEST_F(DescriptorTest, check_lower_upper_bounds) {
 
 TEST_F(DescriptorTest, check_lt_eq) {
   EXPECT_EQ(descriptors_[3].integer_range.at(0).from_value,
-            std::numeric_limits<int>::lowest());
+            std::numeric_limits<int64_t>::lowest());
   EXPECT_EQ(descriptors_[3].integer_range.at(0).to_value, 15);
 }
 
 TEST_F(DescriptorTest, check_gt) {
   EXPECT_EQ(descriptors_[4].integer_range.at(0).from_value, 15);
   EXPECT_EQ(descriptors_[4].integer_range.at(0).to_value,
-            std::numeric_limits<int>::max());
+            std::numeric_limits<int64_t>::max());
 }
 
 int main(int argc, char** argv) {
